fix POST & PUT validate, add controllers+shemaJoi
