+++ controllers/contactList.js
@@ -1,0 +1,12 @@
+const contacts = require('../models/contacts')
+
+const contactList =  async (req, res, next) => {
+    try {
+      const result = await contacts.listContacts();
+      res.status(200).json(result);
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  module.exports = contactList
