+++ routes/api/contacts.js
@@ -1,87 +1,22 @@
 
 const express = require("express");
 const router = express.Router();
-const contacts = require("../../models/contacts");
-const Joi = require("joi");
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
+const findContactById = require("../../controllers/findContactById");
+const addContact = require("../../controllers/addContact");
+const deleteContact = require("../../controllers/deleteContact");
+const updateContact = require("../../controllers/updateContact");
+const contactList = require("../../controllers/contactList");
 
-router.get("/", async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.getContactById(contactId);
+router.get("/", contactList);
 
-    if (!result) {
-      throw res.status(404).json({ message: "Not found" });
-    }
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", findContactById);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if (error) {
-      throw res.status(400).json({ message: "missing required name field" });
-    }
-    const result = await contacts.addContact(req.body);
-    if (!result) {
-      throw res.status(404).json({ message: "Not found" });
-    }
-    console.log("Contact added! New lists of contacts");
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", addContact);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.removeContact(contactId);
-    if (!result) {
-      throw res.status(404).json({ message: "Not found" });
-    }
-    res.status(200).json({ message: "contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", deleteContact);
 
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = addSchema.validate(req.body);
-    if(req.body.constructor === Object && Object.keys(req.body).length === 0) {
-      throw res.status(400).json({ message: "missing fields" });
-    }
-
-    if (error) {
-      throw res.status(400).json({message: "Please fill out all fields"});
-    }
-    const { contactId } = req.params;
-    const result = await contacts.updateContact(contactId, req.body);
-    if (!result) {
-      throw res.status(404).json({ message: "Not found" });
-    }
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:contactId", updateContact);
 
 module.exports = router;
